+++ packages/expo-auth-session/src/TokenRequest.ts
@@ -157,6 +157,23 @@ export class Request<T, B> {
     throw new Error('getQueryBody must be extended');
   }
 }
+function sanitizeExtraHeaders(
+  extra: Record<string, string> | undefined,
+  hasClientSecret: boolean
+): Record<string, string> | undefined {
+  if (!extra) {
+    return undefined;
+  }
+  const extraHeaders = { ...extra };
+  delete extraHeaders['Content-Type'];
+  delete extraHeaders['content-type'];
+  if (hasClientSecret) {
+    // auth-session will set the Authorization header in this case
+    delete extraHeaders.authorization;
+    delete extraHeaders.Authorization;
+  }
+  return extraHeaders;
+}
 
 /**
  * A generic token request.
@@ -179,12 +196,17 @@ export class TokenRequest<T extends TokenRequestConfig>
     this.clientId = request.clientId;
     this.clientSecret = request.clientSecret;
     this.extraParams = request.extraParams;
-    this.extraHeaders = request.extraHeaders;
     this.scopes = request.scopes;
+    this.extraHeaders = sanitizeExtraHeaders(
+      request.extraHeaders,
+      typeof request.clientSecret !== 'undefined'
+    );
   }
 
   getHeaders(): Headers {
-    const headers: Headers = { 'Content-Type': 'application/x-www-form-urlencoded' };
+    const headers: Headers = Object.assign({}, this.extraHeaders, {
+      'Content-Type': 'application/x-www-form-urlencoded',
+    });
     if (typeof this.clientSecret !== 'undefined') {
       // If client secret exists, it should be converted to base64
       // https://tools.ietf.org/html/rfc6749#section-2.3.1
@@ -193,14 +215,6 @@ export class TokenRequest<T extends TokenRequestConfig>
       const credentials = `${encodedClientId}:${encodedClientSecret}`;
       const basicAuth = Base64.encodeNoWrap(credentials);
       headers.Authorization = `Basic ${basicAuth}`;
-    }
-
-    if (this.extraHeaders) {
-      for (const extra in this.extraHeaders) {
-        if (extra in this.extraHeaders && !(extra in headers)) {
-          headers[extra] = this.extraHeaders[extra];
-        }
-      }
     }
 
     return headers;
@@ -378,6 +392,7 @@ export class RevokeTokenRequest
   readonly clientSecret?: string;
   readonly token: string;
   readonly tokenTypeHint?: TokenTypeHint;
+  readonly extraHeaders?: Record<string, string>;
 
   constructor(request: RevokeTokenRequestConfig) {
     super(request);
@@ -386,10 +401,16 @@ export class RevokeTokenRequest
     this.clientSecret = request.clientSecret;
     this.token = request.token;
     this.tokenTypeHint = request.tokenTypeHint;
+    this.extraHeaders = sanitizeExtraHeaders(
+      request.extraHeaders,
+      typeof request.clientSecret !== 'undefined'
+    );
   }
 
   getHeaders(): Headers {
-    const headers: Headers = { 'Content-Type': 'application/x-www-form-urlencoded' };
+    const headers: Headers = Object.assign({}, this.extraHeaders, {
+      'Content-Type': 'application/x-www-form-urlencoded',
+    });
     if (typeof this.clientSecret !== 'undefined' && this.clientId) {
       // If client secret exists, it should be converted to base64
       // https://tools.ietf.org/html/rfc6749#section-2.3.1
@@ -428,6 +449,7 @@ export class RevokeTokenRequest
       clientSecret: this.clientSecret,
       token: this.token,
       tokenTypeHint: this.tokenTypeHint,
+      extraHeaders: this.extraHeaders,
     };
   }